+++ darkside-tests/tests/tests.rs
@@ -3,11 +3,11 @@ use darkside_tests::utils::prepare_darksidewalletd;
 use darkside_tests::utils::update_tree_states_for_transaction;
 use tempfile::TempDir;
 use zcash_local_net::indexer::Indexer;
-use zcash_local_net::network::localhost_uri;
-use zingo_common_components::protocol::activation_heights::for_test::all_height_one_nus;
+use zingo_common_components::protocol::ActivationHeights;
 use zingo_test_vectors::seeds::DARKSIDE_SEED;
 use zingolib::get_base_address_macro;
 use zingolib::testutils::lightclient::from_inputs;
+use zingolib::testutils::port_to_localhost_uri;
 use zingolib::testutils::tempfile;
 use zingolib::wallet::balance::AccountBalance;
 use zingolib_testutils::scenarios::ClientBuilder;
@@ -19,11 +19,11 @@ use darkside_tests::utils::lightwalletd;
 async fn simple_sync() {
     let lightwalletd = lightwalletd().await.unwrap();
 
-    let server_id = localhost_uri(lightwalletd.listen_port());
+    let server_id = port_to_localhost_uri(lightwalletd.listen_port());
     prepare_darksidewalletd(server_id.clone(), true)
         .await
         .unwrap();
-    let activation_heights = all_height_one_nus();
+    let activation_heights = ActivationHeights::default();
     let wallet_dir = TempDir::new().unwrap();
     let mut light_client = ClientBuilder::new(server_id, wallet_dir).build_client(
         DARKSIDE_SEED.to_string(),
@@ -62,12 +62,12 @@ async fn simple_sync() {
 async fn reorg_receipt_sync_generic() {
     let lightwalletd = lightwalletd().await.unwrap();
 
-    let server_id = localhost_uri(lightwalletd.listen_port());
+    let server_id = port_to_localhost_uri(lightwalletd.listen_port());
     prepare_darksidewalletd(server_id.clone(), true)
         .await
         .unwrap();
 
-    let activation_heights = all_height_one_nus();
+    let activation_heights = ActivationHeights::default();
     let wallet_dir = TempDir::new().unwrap();
     let mut light_client = ClientBuilder::new(server_id.clone(), wallet_dir).build_client(
         DARKSIDE_SEED.to_string(),
@@ -122,14 +122,14 @@ async fn reorg_receipt_sync_generic() {
 async fn sent_transaction_reorged_into_mempool() {
     let lightwalletd = lightwalletd().await.unwrap();
 
-    let server_id = localhost_uri(lightwalletd.listen_port());
+    let server_id = port_to_localhost_uri(lightwalletd.listen_port());
     prepare_darksidewalletd(server_id.clone(), true)
         .await
         .unwrap();
 
     let wallet_dir = TempDir::new().unwrap();
     let mut client_manager = ClientBuilder::new(server_id.clone(), wallet_dir);
-    let activation_heights = all_height_one_nus();
+    let activation_heights = ActivationHeights::default();
     let mut light_client =
         client_manager.build_client(DARKSIDE_SEED.to_string(), 0, true, activation_heights);
     let mut recipient = client_manager.build_client(