+++ zingolib/src/lightclient/propose.rs
@@ -17,7 +17,7 @@ use crate::wallet::error::ProposeShieldError;
 
 impl LightClient {
     fn append_zingo_zenny_receiver(&self, receivers: &mut Vec<Receiver>) {
-        let zfz_address = get_donation_address_for_chain(&self.config().chain);
+        let zfz_address = get_donation_address_for_chain(&self.config().network_type());
         let dev_donation_receiver = Receiver::new(
             crate::utils::conversion::address_from_str(zfz_address).expect("Hard coded str"),
             Zatoshis::from_u64(ZENNIES_FOR_ZINGO_AMOUNT).expect("Hard coded u64."),
@@ -182,10 +182,10 @@ mod shielding {
     };
 
     fn create_basic_client() -> LightClient {
-        let config = ZingoConfigBuilder::default().create();
+        let config = ZingoConfigBuilder::default().build();
         LightClient::create_from_wallet(
             LightWallet::new(
-                config.chain,
+                config.network_type(),
                 WalletBase::Mnemonic {
                     mnemonic: Mnemonic::from_phrase(seeds::HOSPITAL_MUSEUM_SEED.to_string())
                         .unwrap(),