+++ zingo-cli/CHANGELOG.md
@@ -12,11 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
-`remove_transaction` command - now only allows transactions with the new `Failed` status to be removed.
+- `remove_transaction` command: now only allows transactions with the new `Failed` status to be removed.
 
 ### Removed
-`resend` command - see zingolib CHANGELOG.md on `LightClient::resend`
-`send_progress` command
+- `resend` command: see zingolib CHANGELOG.md on `LightClient::resend`
+- `send_progress` command
+- `regtest` feature: can still use zingo-cli in regtest mode with no features enabled using the '--chain regtest' flag. 
 
 ## [0.2.0]