+++ zingo-cli/CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+
+### Removed
+
+## [0.3.0] - 2026-06-05
+
+### Changed
 `remove_transaction` command - now only allows transactions with the new `Failed` status to be removed.
 
 ### Removed